move replace functions to helper module
